% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{as_integrations}
\alias{as_integrations}
\title{Coerce to integration records}
\usage{
as_integrations(x)
}
\arguments{
\item{x}{A data frame, file path, or \code{vi_integrations} object.}
}
\value{
A \code{vi_integrations} object.
}
\description{
Coerce to integration records
}
