% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_genome.R
\name{create_genome_config}
\alias{create_genome_config}
\title{Create Genome Configuration for Circos}
\usage{
create_genome_config(
  host_file,
  virus_name,
  virus_length,
  virus_visual_ratio = 0.1,
  color_file = NULL
)
}
\arguments{
\item{host_file}{Path to chromosome sizes file.}

\item{virus_name}{Name of the virus (e.g., "HBV").}

\item{virus_length}{Length of the viral genome in bp.}

\item{virus_visual_ratio}{The ratio of the circle occupied by the virus.}

\item{color_file}{Optional path to a CSV with chromosome colors.}
}
\description{
Create Genome Configuration for Circos
}
