% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{create_gi}
\alias{create_gi}
\title{Create GInteractions Object from VCF}
\usage{
create_gi(vcf_file)
}
\arguments{
\item{vcf_file}{Path to the VCF file.}
}
\value{
A GInteractions object.
}
\description{
Create GInteractions Object from VCF
}
\keyword{internal}
