% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{create_gi_from_table}
\alias{create_gi_from_table}
\title{Convert an Integration Table to GInteractions}
\usage{
create_gi_from_table(input_file, cfg)
}
\arguments{
\item{input_file}{Path to an integration table.}

\item{cfg}{Plotting configuration returned by \code{create_config()}.}
}
\value{
A \code{GInteractions} object.
}
\description{
Convert an Integration Table to GInteractions
}
