% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_circos.R
\name{create_global_style}
\alias{create_global_style}
\title{Global Style Configuration for ViHostPlot}
\usage{
create_global_style(...)
}
\arguments{
\item{...}{Style parameters to override defaults.}
}
\description{
Global Style Configuration for ViHostPlot
}
