% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{draw_histogram}
\alias{draw_histogram}
\title{Draw Histogram Track in Genomic Mode}
\usage{
draw_histogram(
  data,
  height,
  cfg,
  track_label = NULL,
  bins = NULL,
  col = "#4D4D4D"
)
}
\arguments{
\item{data}{Data frame converted from a \code{GInteractions} object.}

\item{height}{Numeric. Track height.}

\item{cfg}{Plotting configuration list.}

\item{track_label}{Optional track label to draw on the HBV sector.}

\item{bins}{Number of genomic bins per sector.}

\item{col}{Histogram fill color.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draw Histogram Track in Genomic Mode
}
