% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{draw_scatter}
\alias{draw_scatter}
\title{Draw Scatter Track in Genomic Mode}
\usage{
draw_scatter(
  data,
  height,
  cfg,
  track_label = NULL,
  track_label_cex = 0.8,
  track_label_col = "grey30",
  method_col = NULL,
  point_color = "blue",
  baseline_col = "grey90"
)
}
\arguments{
\item{data}{Data frame converted from a \code{GInteractions} object.}

\item{height}{Numeric. Track height.}

\item{cfg}{Plotting configuration list.}

\item{track_label}{Optional track label to draw on the virus sector.}

\item{track_label_cex}{Track label text size.}

\item{track_label_col}{Track label color.}

\item{method_col}{Optional named vector of colors for different methods.}

\item{point_color}{Default point color.}

\item{baseline_col}{Baseline color for the scatter track.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draw Scatter Track in Genomic Mode
}
