% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{get_virus_name}
\alias{get_virus_name}
\title{Detect Virus Name from Chromosome Sizes
logic: Exclusion of mitochondria, then finding the shortest sequence.}
\usage{
get_virus_name(chrom_sizes_df)
}
\arguments{
\item{chrom_sizes_df}{Data frame with 'chr' and 'end'.}
}
\value{
A list containing 'name' and 'length' of the virus.
}
\description{
Detect Virus Name from Chromosome Sizes
logic: Exclusion of mitochondria, then finding the shortest sequence.
}
\keyword{internal}
