% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_features.R
\name{hbv_features}
\alias{hbv_features}
\title{Built-in HBV feature annotations}
\usage{
hbv_features(version = "ayw")
}
\arguments{
\item{version}{HBV annotation version. Currently only \code{"ayw"} is supported.}
}
\value{
A \code{vi_virus_features} object.
}
\description{
Built-in HBV feature annotations
}
