% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_chrom_utils.R
\name{host_genome}
\alias{host_genome}
\title{Construct a host genome table}
\usage{
host_genome(host = "hg38", chrom_file = NULL, colors = NULL)
}
\arguments{
\item{host}{Built-in host name, chromosome table, or \code{NULL} when
\code{chrom_file} is supplied.}

\item{chrom_file}{Optional path to a chromosome-size table.}

\item{colors}{Optional named vector of chromosome colors.}
}
\value{
A host genome object with class \code{vi_host_genome}.
}
\description{
Construct a host genome table
}
