% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{plan_layout}
\alias{plan_layout}
\title{Plan Track Layout}
\usage{
plan_layout(labels, space)
}
\arguments{
\item{labels}{Vector of sample labels.}

\item{space}{Total available space.}
}
\value{
Named vector of track heights.
}
\description{
Plan Track Layout
}
\keyword{internal}
