% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plot_oncoprint}
\alias{plot_oncoprint}
\title{Plot oncoprint with annotation tracks}
\usage{
plot_oncoprint(
  maf,
  genes = 20,
  annotations = NULL,
  numeric_bins = 5,
  annotation_colors = NULL,
  sort_by = NULL,
  annotation_levels = NULL
)
}
\arguments{
\item{maf}{MAF object.}

\item{genes}{The gene names or the number of genes to display.}

\item{annotations}{Character vector or list of annotation variables.}

\item{numeric_bins}{Number of bins for continuous annotation variables.}

\item{annotation_colors}{Named list of colors for annotations.}

\item{sort_by}{Annotation variable used to sort samples.}

\item{annotation_levels}{Named list of explicit level orders.}
}
\value{
A combined plot object.
}
\description{
Plot oncoprint with annotation tracks
}
