% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_circos.R
\name{plot_virus_integration}
\alias{plot_virus_integration}
\title{Draw Viral Integration Tracks}
\usage{
plot_virus_integration(gi, config, style = create_global_style())
}
\arguments{
\item{gi}{Normalized GenomicInteractions object.}

\item{config}{GenomeConfig object.}

\item{style}{Global style list.}
}
\description{
Draw Viral Integration Tracks
}
