% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.vi_integrations}
\alias{print.vi_integrations}
\title{Print integration records}
\usage{
\method{print}{vi_integrations}(x, ...)
}
\arguments{
\item{x}{A \code{vi_integrations} object.}

\item{...}{Ignored.}
}
\description{
Print integration records
}
