% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{raw_to_vcf_simple}
\alias{raw_to_vcf_simple}
\title{Convert Raw Table to Temporary VCF}
\usage{
raw_to_vcf_simple(input_file, output_vcf, host_file, virus_info)
}
\arguments{
\item{input_file}{Path to input tabular data.}

\item{output_vcf}{Path to output temporary VCF.}

\item{host_file}{Path to host chromosome sizes file.}

\item{virus_info}{List containing virus name and length.}
}
\description{
Convert Raw Table to Temporary VCF
}
\keyword{internal}
