% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{read_integrations}
\alias{read_integrations}
\title{Read integration records}
\usage{
read_integrations(path, sep = "\\t")
}
\arguments{
\item{path}{Path to a tabular file containing integration records.}

\item{sep}{Field separator. Defaults to tab.}
}
\value{
A \code{vi_integrations} object.
}
\description{
Read integration records
}
