% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{track_links}
\alias{track_links}
\title{Create a link track specification}
\usage{
track_links(
  color = NULL,
  radius = NULL,
  lwd = 0.35,
  default_col = "grey",
  subset = NULL,
  method_col = NULL
)
}
\arguments{
\item{color}{Column name or literal color for links.}

\item{radius}{Optional link radius.}

\item{lwd}{Link width.}

\item{default_col}{Default link color.}

\item{subset}{Optional named list of filters.}

\item{method_col}{Optional named vector of colors for explicit mapping.}
}
\value{
A track object for \code{plot_integrations()}.
}
\description{
Create a link track specification
}
