% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{track_virus_genes}
\alias{track_virus_genes}
\title{Create a virus feature track specification}
\usage{
track_virus_genes(
  features,
  height = 0.1,
  fill = NULL,
  label = TRUE,
  label_cex = 0.32,
  label_col = "grey20",
  label_min_width = 250,
  border_col = "white"
)
}
\arguments{
\item{features}{Virus feature intervals from \code{virus_features()}.}

\item{height}{Track height.}

\item{fill}{Feature fill color or named vector by feature type.}

\item{label}{Logical. Whether to label features.}

\item{label_cex}{Feature label text size.}

\item{label_col}{Feature label color.}

\item{label_min_width}{Minimum feature width, in base pairs, required for
drawing an internal feature label.}

\item{border_col}{Feature border color.}
}
\value{
A track object for \code{plot_integrations()}.
}
\description{
Create a virus feature track specification
}
