% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_integrations}
\alias{validate_integrations}
\title{Validate integration records}
\usage{
validate_integrations(x)
}
\arguments{
\item{x}{Integration records.}
}
\value{
A standardized \code{vi_integrations} object.
}
\description{
Validate integration records
}
