% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_chrom_utils.R
\name{virus_genome}
\alias{virus_genome}
\title{Construct a virus genome table}
\usage{
virus_genome(name, length = NULL)
}
\arguments{
\item{name}{Virus sequence name, or a named numeric vector of length 1.}

\item{length}{Virus sequence length in base pairs.}
}
\value{
A virus genome object with class \code{vi_virus_genome}.
}
\description{
Construct a virus genome table
}
