% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{create_config}
\alias{create_config}
\title{Create Plotting Configuration}
\usage{
create_config(host_file, virus_info, visual_ratio, color_file)
}
\arguments{
\item{host_file}{Path to host file.}

\item{virus_info}{List of virus name and length.}

\item{visual_ratio}{Visual ratio for the virus sector.}

\item{color_file}{Optional path to color CSV.}
}
\value{
A list containing plotting configuration.
}
\description{
Create Plotting Configuration
}
\keyword{internal}
