% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{draw_ideogram}
\alias{draw_ideogram}
\title{Draw Ideogram Track}
\usage{
draw_ideogram(height, cfg)
}
\arguments{
\item{height}{Numeric. The height of the track}

\item{cfg}{List. The configuration list containing \code{grid_col} (colors).}
}
\description{
Draw Ideogram Track
}
