% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{draw_ideogram}
\alias{draw_ideogram}
\title{Draw Ideogram Track in Genomic Mode}
\usage{
draw_ideogram(height, cfg)
}
\arguments{
\item{height}{Numeric. Track height.}

\item{cfg}{Plotting configuration list.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draw Ideogram Track in Genomic Mode
}
