% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos2.R
\name{draw_ideogram}
\alias{draw_ideogram}
\title{Draw Ideogram Track in Genomic Mode}
\usage{
draw_ideogram(height, cfg, grid_col = NULL, border_col = "white")
}
\arguments{
\item{height}{Numeric. Track height.}

\item{cfg}{Plotting configuration list.}

\item{grid_col}{Optional named vector of sector colors.}

\item{border_col}{Border color for ideogram rectangles.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draw Ideogram Track in Genomic Mode
}
