% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{draw_link}
\alias{draw_link}
\title{Draw Link Layer in Genomic Mode}
\usage{
draw_link(link_data, cfg, radius = NULL)
}
\arguments{
\item{link_data}{Data frame converted from a \code{GInteractions} object.}

\item{cfg}{Plotting configuration list.}

\item{radius}{Optional link radius.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draw Link Layer in Genomic Mode
}
