% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{draw_link}
\alias{draw_link}
\title{Draw Link Layer (Arcs)}
\usage{
draw_link(link_data, cfg, radius = NULL)
}
\arguments{
\item{link_data}{Data frame. Must contain columns: seqnames1, start1, seqnames2, start2, Source(optional).}

\item{cfg}{List. Configuration list containing data about chromosomes.}

\item{radius}{Numeric (Optional). The radius to draw links on. If NULL, defaults to the bottom of the current track.}
}
\description{
Draw Link Layer (Arcs)
}
