% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{draw_scatter}
\alias{draw_scatter}
\title{Draw Scatter Track in Genomic Mode}
\usage{
draw_scatter(data, height, cfg, point_color = "blue")
}
\arguments{
\item{data}{Data frame converted from a \code{GInteractions} object.}

\item{height}{Numeric. Track height.}

\item{cfg}{Plotting configuration list.}

\item{point_color}{Default point color.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draw Scatter Track in Genomic Mode
}
