% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{draw_scatter}
\alias{draw_scatter}
\title{Draw Scatter Plot Track}
\usage{
draw_scatter(data, height, cfg, label_col = "Label", point_color = "blue")
}
\arguments{
\item{data}{Data frame. The full dataset (must be converted from GI object).
Must contain columns: seqnames1, start1, Depth.}

\item{height}{Numeric. The height of the track.}

\item{cfg}{List. Configuration list.}

\item{label_col}{String. The column name in \code{data} to use for identifying the sample.}

\item{point_color}{String. Color of the points. Default is "blue".}
}
\description{
Draw Scatter Plot Track
}
