% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{get_virus_name}
\alias{get_virus_name}
\title{Detect Virus Sequence}
\usage{
get_virus_name(chrom_sizes_df)
}
\arguments{
\item{chrom_sizes_df}{Data frame returned by \code{read_chrom_sizes()}.}
}
\value{
The inferred virus sequence name.
}
\description{
Detect Virus Sequence
}
\keyword{internal}
