% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot}
\alias{oncoplot}
\title{Generate a customized oncoplot with clinical tracks}
\usage{
oncoplot(maf, genes = 20, clinical_data = NULL, clinical_vars = NULL)
}
\arguments{
\item{maf}{A MAF object generated by read.maf}

\item{genes}{Number of top genes to plot, default 20}

\item{clinical_data}{Data frame containing clinical information}

\item{clinical_vars}{List defining variables and their types (discrete/continuous)}
}
\value{
A patchwork/aplot object
}
\description{
Generate a customized oncoplot with clinical tracks
}
