% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot}
\alias{oncoplot}
\title{Generate a oncoplot with clinical tracks}
\usage{
oncoplot(
  maf,
  genes = 20,
  clinical_vars = NULL,
  n_breaks = 5,
  colors = NULL,
  clinical_palettes = NULL
)
}
\arguments{
\item{maf}{MAF object.}

\item{genes}{the gene names or the number, default is 20.}

\item{clinical_vars}{A character vector or list defining clinical variables to display as bottom tracks.}

\item{n_breaks}{Integer. Number of breaks for continuous clinical variables. Default is 5.}

\item{colors}{A named list of customized colors for clinical tracks.}

\item{clinical_palettes}{A named list or character vector of RColorBrewer palettes for clinical tracks.}
}
\value{
A combined plot object (class depends on \code{aplot} output).
}
\description{
Generate a oncoplot with clinical tracks
}
\examples{
\donttest{
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml.clin <- system.file('extdata', 'tcga_laml_annot.tsv', package = 'maftools')
laml <- maftools::read.maf(maf = laml.maf, clinicalData = laml.clin)
var_names <- c("FAB_classification", "days_to_last_followup")
oncoplot(maf = laml, genes = 20, clinical_vars = var_names)
}
}
