% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{track_ideogram}
\alias{track_ideogram}
\title{Create an ideogram track specification}
\usage{
track_ideogram(height = 0.08, grid_col = NULL, border_col = "white")
}
\arguments{
\item{height}{Track height.}

\item{grid_col}{Optional named vector of chromosome colors.}

\item{border_col}{Border color for the ideogram rectangles.}
}
\value{
A track object for \code{plot_integrations()}.
}
\description{
Create an ideogram track specification
}
