% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{track_virus_genes}
\alias{track_virus_genes}
\title{Create a virus feature track specification}
\usage{
track_virus_genes(
  features,
  height = 0.08,
  fill = NULL,
  label = TRUE,
  border_col = "white"
)
}
\arguments{
\item{features}{Virus feature intervals from \code{virus_features()}.}

\item{height}{Track height.}

\item{fill}{Feature fill color or named vector by feature type.}

\item{label}{Logical. Whether to label features.}

\item{border_col}{Feature border color.}
}
\value{
A track object for \code{plot_integrations()}.
}
\description{
Create a virus feature track specification
}
